+++ CHANGELOG.md
@@ -5,3 +5,4 @@ All notable changes to the "favorite-git-commands" extension will be documented 
 ## [0.0.10] - 2025-12-06
 - Replaced "Paste Command" with "Copy to Clipboard" functionality.
 - Added "Move Command" functionality to easily move commands between folders.
+- Renamed extension and commands from `favorite-git-*` to `favorite-cli-*` to reflect broader usage.